feat: add main file fo test initialization
